chore(papyrus_base_layer): simplify price sample creation post alloy version update

+++ crates/papyrus_base_layer/src/lib.rs
@@ -70,7 +70,7 @@ pub trait BaseLayerContract {
 /// A struct that holds together the data on the base layer's gas prices, for a given timestamp.
 pub struct PriceSample {
     pub timestamp: u64,
-    pub base_fee_per_gas: u128,
+    pub base_fee_per_gas: u64,
     pub blob_fee: u128,
 }
